requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.3
numpy==2.1.1
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==2.6.1
torch>=2.5.0
transformers==4.35.2
tokenizers==0.15.0
python-dotenv==1.0.1
httpx[http2]==0.27.2